FastAPI/v5/fastapi-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
pydantic[email]==1.10.13
python-dotenv==1.0.0
httpx==0.24.1
#grpcio==1.60.0
#grpcio-tools==1.60.0
pytest==7.4.3
pytest-asyncio==0.21.1
authlib==1.2.1
python-jose[cryptography]==3.3.0
#protobuf>=4.21.12
grpcio==1.73.1
grpcio-tools==1.73.1
protobuf==6.31.0